moved more helper functions to helper_functions file
